# Conflicts:
#	src/main/java/switch2021/project/controller/RefineUserStoryController.java

+++ src/main/java/switch2021/project/controller/ChangePriorityUSController.java
@@ -51,7 +51,7 @@ public class ChangePriorityUSController {
         return this.userStoryList;
     }
 
-    public UserStory getUS(long idUS){
+    public UserStory getUS(int idUS){
         this.userStory = this.productBacklog.getUserStoryById(idUS);
         return this.userStory;
     }
